Make the build reproducible
Whilst working on the Reproducible Builds effort [0], we noticed
that node-promise could not be built reproducibly as it uses
random numbers as throwaway identifiers.

This patch uses a determinstic name for these identifiers based on the
contents of the src/ directory.

This was filed in @Debian as https://bugs.debian.org/886277. There was
also a previous aborted attempt in #146.

 [0] https://reproducible-builds.org/
 [1] https://github.com/then/promise/pull/146

Signed-off-by: Chris Lamb <chris@chris-lamb.co.uk>